[torchsim] The queues carry 32 bits, and nothing here reads what they mean
`std::queue<float>` was the last place this unit had an opinion about types, and
it has had no use for one since the reduce retired. It moves bits between lanes;
a float, an integer, a NaN and a lane number are all just words to it.

    in/out           std::queue<float>  ->  std::queue<uint32_t>
    push at e8/e16   converted (int8 -> float, fp16 -> fp32)  ->  the bits, in the
                     low end of a word
    push at e32      reinterpreted  ->  read as what it is
    permute pattern  arrived as a float spelling of a lane number and was cast
                     back  ->  arrives as the number

THE PATTERN IS THE HALF THAT HAD TO MOVE WITH THE COMPILER. lower_to_vcix built
the lane index with SIToFP and this file read it back with (uint32_t); both go in
the same change, because either alone is silently wrong -- an integer 3 read as a
float gives 4.2e-45 and casts to 0, so every lane would read lane 0, and a float
3.0 read as an integer gives 0x40400000, which is past the end, so every lane
would read nothing.

The e8/e16 meaning changes with it: those paths CONVERTED and now MOVE. That is
what a unit doing no arithmetic should do, and nothing reaches them today --
lower_to_vcix refuses anything narrower than 32 bits before it gets here.

The unit's own checks carry it: `test_bits_survive` no longer smuggles its
patterns through a float to state them.

481/481 on the kernel suite with the compiler's half, no verdict moved.

+++ riscv/cross_lane_unit.h
@@ -1,7 +1,9 @@
 // The cross-lane unit: a tile goes in a vector register at a time and comes back
 // with the LANE AXIS rewritten -- transposed, replicated or permuted. IT DOES NO
 // ARITHMETIC: a reduction is a transpose and then the VECTOR UNIT's own fold, which
-// is the half that knows the type. ONE QUEUE PER
+// is the half that knows the type. SO THE QUEUES CARRY 32 RAW BITS and nothing here
+// reads what they mean -- an integer, a float, a NaN and a lane number all cross as
+// the bits they are. ONE QUEUE PER
 // LANE, like systolicArray_t; the push IS the serialisation and the op runs between.
 #ifndef _RISCV_CROSS_LANE_UNIT_H
 #define _RISCV_CROSS_LANE_UNIT_H
@@ -32,8 +34,8 @@ public:
   uint32_t n_lane;
   // The tile on the way in and the tile on the way out. `in[L]` is lane L's row as
   // it was pushed; `out[L]` is lane L's column once `transpose` has run.
-  std::queue<float> **in;
-  std::queue<float> **out;
+  std::queue<uint32_t> **in;
+  std::queue<uint32_t> **out;
   // How deep each lane's row is -- the number of values pushed per lane since the
   // last transpose. IT IS THE TILE'S OTHER DIMENSION, so the transpose reads it
   // rather than assuming the tile is square.
@@ -45,9 +47,9 @@ public:
 public:
   void reset();
   void run();
-  void transpose(const std::vector<std::vector<float> > &tile);
-  void broadcast(const std::vector<std::vector<float> > &tile);
-  void permute(const std::vector<std::vector<float> > &tile);
+  void transpose(const std::vector<std::vector<uint32_t> > &tile);
+  void broadcast(const std::vector<std::vector<uint32_t> > &tile);
+  void permute(const std::vector<std::vector<uint32_t> > &tile);
 
   crossLaneUnit_t(processor_t *p, reg_t n_vu) : p(p),
                                                 n_lane(n_vu),
@@ -79,7 +81,7 @@ public:
     }
   }
 
-  void push(uint32_t lane, float val)
+  void push(uint32_t lane, uint32_t val)
   {
     in[lane]->push(val);
   }
@@ -89,9 +91,9 @@ public:
     op = o;
   }
 
-  float pop(uint32_t lane)
+  uint32_t pop(uint32_t lane)
   {
-    float val = out[lane]->front();
+    uint32_t val = out[lane]->front();
     out[lane]->pop();
     return val;
   }
